# docs/openapi/operator-v1.yaml
openapi: 3.1.0
info:
  title: HopClaw Operator API
  version: '1.0'
  description: Machine-readable OpenAPI document for the current HopClaw operator and control-plane surface under /operator/*. This API evolves faster than the runtime API, so nested payloads are intentionally modeled as open objects unless they are already stable and obvious from the live handlers.
servers:
- url: http://127.0.0.1:16280
  description: Local default server
security:
- bearerAuth: []
...
paths:
...
  /operator/hooks:
    get:
      tags:
      - Hooks
      operationId: getHooks
      summary: Get hooks
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountedItemsResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - Hooks
      operationId: postHooks
      summary: Post hooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenObject'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /operator/hooks/{id}:
    patch:
      tags:
      - Hooks
      operationId: patchHooksId
      summary: Patch hooks id
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Path parameter id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenObject'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - Hooks
      operationId: deleteHooksId
      summary: Delete hooks id
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Path parameter id.
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenObject'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      additionalProperties: true
    OpenObject:
      type: object
      description: Open JSON object used for faster-moving operator payloads.
      additionalProperties: true
    OpenValue:
      description: Any JSON value accepted or returned by operator config mutation endpoints.
      oneOf:
      - type: object
        additionalProperties: true
      - type: array
        items:
          $ref: '#/components/schemas/OpenValue'
      - type: string
      - type: integer
      - type: number
      - type: boolean
      - type: 'null'
    CountedItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OpenValue'
        count:
          type: integer
      additionalProperties: true
...